Added simulation for all GPIO pins and added buttons for input (not hooked up yet)

+++ mac/m8rsim/Device.h
@@ -19,7 +19,7 @@ typedef NSMutableArray<NSDictionary*>* FileList;
 - (void)addDevice:(NSString*)name;
 - (void)setDevice:(NSString*)name;
 - (void)setContents:(NSData*)contents withName:(NSString*)name;
-- (void)updateGPIOState:(uint16_t) state withMode:(uint16_t) mode;
+- (void)updateGPIOState:(uint32_t) state withMode:(uint32_t) mode;
 - (void)outputMessage:(NSString*) message toBuild:(BOOL) build;
 - (void)reloadFiles;
 - (void) setFreeMemory:(NSUInteger)size numObj:(NSUInteger)obj numStr:(NSUInteger)str numOther:(NSUInteger)oth;
