+++ mac/m8rsim/Device.mm
@@ -45,7 +45,7 @@ class MyGPIOInterface;
 }
 
 - (void)outputMessage:(NSString*)msg, ...;
-- (void)updateGPIOState:(uint16_t) state withMode:(uint16_t) mode;
+- (void)updateGPIOState:(uint32_t) state withMode:(uint32_t) mode;
 - (void)updateMemoryInfo;
 
 @end
@@ -130,7 +130,7 @@ private:
     });
 }
 
-- (void)updateGPIOState:(uint16_t) state withMode:(uint16_t) mode
+- (void)updateGPIOState:(uint32_t) state withMode:(uint32_t) mode
 {
     dispatch_async(dispatch_get_main_queue(), ^{
         [self.delegate updateGPIOState:state withMode:mode];
