+++ mac/m8rsim/SimulationView.h
@@ -12,7 +12,7 @@
 
 @interface SimulationView : NSViewController
 
-- (void)updateGPIOState:(uint16_t) state withMode:(uint16_t) mode;
+- (void)updateGPIOState:(uint32_t) state withMode:(uint32_t) mode;
 
 @end
